move to gdt-dev/core switch
Switches the import of core functionality to the gdt-dev/core
repository.

Issue gdt-dev/gdt#55
