+++ fixtures/kind/kind_test.go
@@ -12,8 +12,8 @@ import (
 	"path/filepath"
 	"testing"
 
+	gdtcontext "github.com/gdt-dev/core/context"
 	"github.com/gdt-dev/gdt"
-	gdtcontext "github.com/gdt-dev/gdt/context"
 	kindfix "github.com/gdt-dev/kube/fixtures/kind"
 	"github.com/stretchr/testify/require"
 )
